Fix large population leakage bug.
Still not sure what the issue was... when there are a lot of particles, removing the reflective BC tanks performance, and the particle population kept increasing. Maybe something to do with modifying the collection while holding references to elements (pretty sure i wasnt doing that), or the particle removal (though i tried a more naive approach copying to new vector, and that didnt fix it either). In the end, I made a back vector that was populated with each particle as it was moved or processed or born from fission as needed, then swap at the end.